Only convert N to false if we're emulating booleans
If a table with a Y/N string boolean column has a default of 'N', then
we do not want to convert that to false. An example

//schema
create table :user do |t|
 t.string "active", limit: 1, default: "N"
end

//console
u = User.new.inspect
<User active: "f">

Active ends up as 'f' because this code converts the default "N" to the
primitive false, which is then cast to the string "false" and truncated
down to 1 character so that it fits in the db column.

Now this conversion only happens if the client has turned on the
'emulate_booleans_from_strings' configuration.