Fix undefined method error in schema dumper
Addresses #802

It only fixes what is generating the `undefined method `[]' for
nil:NilClass` error that is getting dumped in the schema.rb file. The
problem is that the `@types` variable was dropped in
[rails commit
9e4ed2f996](https://github.com/rails/rails/commit/9e4ed2f996d8dd7eaac0399f4efb5ce58aa593f8).
Instead of using `@types`, use `@connection.native_database_types`,
which is what `@types` was set to before `@types` was dropped.

Added a helper method `valid_type?` to follow the conventions used by
the other adapters.