Set table_name after super in Column#initialize
As of
[0aa83f3b8](https://github.com/rails/rails/commit/0aa83f3b88c71df4f29b8c52d6297e2e5886e258)
`ActiveRecord::ConnectionAdapters::Column#initialize` sets `@table_name`
to nil, which undoes the setting of `@table_name` in
OracleEnhancedColumn#initialize. The missing `table_name` causes output
like:
```
 # Could not dump table "tags" because of following
 # ActiveRecord::ConnectionAdapters::OracleEnhancedConnectionException
 #   "DESC " failed; does it exist?
```
when the schema dumper checks for column comments.

As of
[6b6a01c](https://github.com/rails/rails/commit/6b6a01c52bacb2df2914029309144eacff25a3de)
`table_name` is an argument to
`ActiveRecord::ConnectionAdapters::Column#initialize` but that is only
on master, not the beta release of ActiveRecord 5. In the meantime, move
the setting of `@table_name` to after the `super` call so the value does
not get overwritten.

This combined with @DebbieGillespie's change allows us to dump the
schema and run the schema_dumper tests.