Remove `lengths`, `where`, `using` from `OracleEnhanced::IndexDefinition` because Oracle database does not support these options for indexes.
Meaning of each option from MySQL and PostgreSQL point of view:

- lengths

https://dev.mysql.com/doc/refman/5.7/en/create-index.html

> For string columns, indexes can be created that use only the leading part of column values, using col_name(length) syntax to specify an index prefix length:

- `where`

https://www.postgresql.org/docs/10/static/sql-createindex.html

> predicate
The constraint expression for a partial index.

- `using

https://dev.mysql.com/doc/refman/5.7/en/create-index.html

> index_type

Some storage engines permit you to specify an index type when creating an index.

Note: `orders` is not used by Oracle enhanced adapter but Oracle database itself supports it.

- orders

https://docs.oracle.com/database/121/SQLRF/statements_5013.htm#SQLRF01209

> ASC | DESC

Use ASC or DESC to indicate whether the index should be created in ascending or descending order. Indexes on character data are created in ascending or descending order of the character values in the database character set.