Add support to change the permissions granted when creating DB
To change the permissions add the following line to the initializer of
the application:
```
 ActiveRecord::ConnectionAdapters::OracleEnhancedAdapter.permissions =
   ["create session", "create table", "create view",
    "create sequence", "create trigger", "ctxapp"]
```

This is useful for when running `rails db:create` from a CI to build a
DB for testing. As it has the credentials and the connection ready, it
would be easier if the required permissions are granted in on go.

For example, when using `add_context_index`, permissions to
`create trigger`, `create procedure`, and `ctxapp` are required. But
these permissions are not included in this db create task.