Allow unmanaged autogenerated sequence names
To still support primary keys through db triggers, without adding back in
support for generating and maintaining them with migrations, add an
configuration boolen:

    unmanaged_autogenerated_sequences = false # default

Which can be used in an initializer file. When set to `true` there will not be
an exception raised when the sequence name is defined in the model with the
previously supported value:

   self.sequence_name = :autogenerated

This option allows for dbs managed outside of the application to use triggers
for primary keys. Without the `autogenerated` option an error is raised when
inserting the record:

    ORA-02289: sequence does not exist