Implement contains in AREL to fix param binding
Active record wants '?' as a bind placeholder. Without this it will not
provide bind variables to exec.

Oracle requires :N style bind parameters, so it a translation is
required.

Something went wrong in the translation between rails 7.1 and rails 7.2
so that the where("CONTAINS(table.col, ?, 0) >0", term) was no longer
getting it's bind parameter translated to ':a1' like it should be.

This change drops down to AREL nodes which ensures that the contains
query is getting passed through the visitor pattern, specifically
hitting the visit_Arel_Nodes_BindParam method to convert the param to
':a1'.

I still do not understand what changed in rails, but this workaround is
effective.