inline quoting of LOBs
I could't find any normal code path where inline quoting of LOBs would
take place. But it might be used by people who generate raw SQL queries
so I think it is worth having it somehow working instead of silently
making it empty data.

For example `#build_fixture_sql` and `#build_fixture_statements` in
ActiveRecord::ConnectionAdapters::DatabaseStatements use it but we
override `#insert_fixture` and `#insert_fixtures_set` which use these.

Actually this `empty_nclob()` does not even work so perhaps this code
path is not used for a very long time. Using it, I get an error

> ORA-00904: "EMPTY_NCLOB": invalid identifier

In the past such fix was rejected #1588 but I think originally this
empty quoting intended to workaround writing the quoted value and then
`write_lob()` method inserting it again. But nowadays `write_lob()` is
actually redundant and removed in another commit.

So basically we have these as a last resort effort to make things work
for niche use cases with the significant limitation on BLOBs length in
particular.